pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promptshield-env"
version = "1.0.0"
description = "Deterministic OpenEnv environment for prompt injection and data exfiltration defense."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.3",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "requests>=2.32.0",
    "websockets>=15.0.1",
    "openai>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.2.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
packages = ["server"]
py-modules = [
    "__init__",
    "app",
    "baseline",
    "client",
    "env",
    "graders",
    "models",
    "reward",
    "tasks",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
